+++ crates/objc2/CHANGELOG.md
@@ -27,6 +27,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   who are `IsAllocableAnyThread`.
 * **BREAKING**: Moved the `MethodImplementation` trait from the `declare`
   module to the `runtime` module.
+* **BREAKING**: Moved the `MessageReceiver` and `MessageArguments` traits to
+  the `runtime` module.
 
 ### Deprecated
 * Soft deprecated using `msg_send!` without a comma between arguments (i.e.
