+++ crates/objc2/src/declare/ivar.rs
@@ -320,8 +320,8 @@ mod tests {
     use crate::declare::{IvarBool, IvarEncode};
     use crate::mutability::Mutable;
     use crate::rc::Id;
-    use crate::runtime::NSObject;
-    use crate::{declare_class, msg_send, msg_send_id, test_utils, ClassType, MessageReceiver};
+    use crate::runtime::{MessageReceiver, NSObject};
+    use crate::{declare_class, msg_send, msg_send_id, test_utils, ClassType};
 
     struct TestIvar;
