+++ crates/test-assembly/crates/test_msg_send_error/expected/apple-aarch64.s
@@ -21,7 +21,7 @@ Lloh1:
 	.loh AdrpAdd	Lloh0, Lloh1
 
 	.p2align	2
-SYM(objc2[CRATE_ID]::message::encountered_error::<objc2[CRATE_ID]::runtime::AnyObject>, 0):
+SYM(objc2[CRATE_ID]::runtime::message::encountered_error::<objc2[CRATE_ID]::runtime::AnyObject>, 0):
 	stp	x29, x30, [sp, #-16]!
 	mov	x29, sp
 	bl	_objc_retain
@@ -59,7 +59,7 @@ _error_bool:
 	ret
 LBB2_2:
 	ldr	x0, [sp, #8]
-	bl	SYM(objc2[CRATE_ID]::message::encountered_error::<objc2[CRATE_ID]::runtime::AnyObject>, 0)
+	bl	SYM(objc2[CRATE_ID]::runtime::message::encountered_error::<objc2[CRATE_ID]::runtime::AnyObject>, 0)
 	ldp	x29, x30, [sp, #16]
 	add	sp, sp, #32
 	ret
@@ -227,31 +227,31 @@ l_anon.[ID].2:
 	.p2align	3, 0x0
 l_anon.[ID].3:
 	.quad	l_anon.[ID].2
-	.asciz	"6\000\000\000\000\000\000\000\013\000\000\000\005\000\000"
+	.asciz	"6\000\000\000\000\000\000\000\n\000\000\000\005\000\000"
 
 	.p2align	3, 0x0
 l_anon.[ID].4:
 	.quad	l_anon.[ID].2
-	.asciz	"6\000\000\000\000\000\000\000\020\000\000\000\005\000\000"
+	.asciz	"6\000\000\000\000\000\000\000\017\000\000\000\005\000\000"
 
 	.p2align	3, 0x0
 l_anon.[ID].5:
 	.quad	l_anon.[ID].2
-	.asciz	"6\000\000\000\000\000\000\000\025\000\000\000\005\000\000"
+	.asciz	"6\000\000\000\000\000\000\000\024\000\000\000\005\000\000"
 
 	.p2align	3, 0x0
 l_anon.[ID].6:
 	.quad	l_anon.[ID].2
-	.asciz	"6\000\000\000\000\000\000\000\032\000\000\000\005\000\000"
+	.asciz	"6\000\000\000\000\000\000\000\031\000\000\000\005\000\000"
 
 	.p2align	3, 0x0
 l_anon.[ID].7:
 	.quad	l_anon.[ID].2
-	.asciz	"6\000\000\000\000\000\000\000\037\000\000\000\005\000\000"
+	.asciz	"6\000\000\000\000\000\000\000\036\000\000\000\005\000\000"
 
 	.p2align	3, 0x0
 l_anon.[ID].8:
 	.quad	l_anon.[ID].2
-	.asciz	"6\000\000\000\000\000\000\000$\000\000\000\005\000\000"
+	.asciz	"6\000\000\000\000\000\000\000#\000\000\000\005\000\000"
 
 .subsections_via_symbols
