+++ crates/test-ui/ui/extern_methods_invalid_type.stderr
@@ -22,7 +22,7 @@ error[E0277]: the trait bound `Id<MyObject>: Encode` is not satisfied
   = note: required for `Id<MyObject>` to implement `EncodeReturn`
   = note: required for `Id<MyObject>` to implement `ConvertReturn`
 note: required by a bound in `send_message`
- --> $WORKSPACE/crates/objc2/src/message.rs
+ --> $WORKSPACE/crates/objc2/src/runtime/message.rs
   |
   |     unsafe fn send_message<A, R>(self, sel: Sel, args: A) -> R
   |               ------------ required by a bound in this associated function
@@ -101,7 +101,7 @@ error[E0277]: the trait bound `Result<(), Id<NSObject>>: Encode` is not satisfie
   = note: required for `Result<(), Id<NSObject>>` to implement `EncodeReturn`
   = note: required for `Result<(), Id<NSObject>>` to implement `ConvertReturn`
 note: required by a bound in `send_message`
- --> $WORKSPACE/crates/objc2/src/message.rs
+ --> $WORKSPACE/crates/objc2/src/runtime/message.rs
   |
   |     unsafe fn send_message<A, R>(self, sel: Sel, args: A) -> R
   |               ------------ required by a bound in this associated function
@@ -157,7 +157,7 @@ error[E0277]: the trait bound `MainThreadMarker: Encode` is not satisfied
   = note: required for `MainThreadMarker` to implement `EncodeReturn`
   = note: required for `MainThreadMarker` to implement `ConvertReturn`
 note: required by a bound in `send_message`
- --> $WORKSPACE/crates/objc2/src/message.rs
+ --> $WORKSPACE/crates/objc2/src/runtime/message.rs
   |
   |     unsafe fn send_message<A, R>(self, sel: Sel, args: A) -> R
   |               ------------ required by a bound in this associated function
