+++ crates/test-ui/ui/msg_send_invalid_error.stderr
@@ -45,7 +45,7 @@ error[E0277]: the trait bound `i32: Message` is not satisfied
             NSCollectionView
           and $N others
 note: required by a bound in `__send_message_error`
- --> $WORKSPACE/crates/objc2/src/message.rs
+ --> $WORKSPACE/crates/objc2/src/runtime/message.rs
   |
   |     unsafe fn __send_message_error<A, E>(self, sel: Sel, args: A) -> Result<(), Id<E>>
   |               -------------------- required by a bound in this associated function
