+++ crates/test-ui/ui/msg_send_not_encode.stderr
@@ -17,7 +17,7 @@ error[E0277]: the trait bound `Vec<u8>: Encode` is not satisfied
   = note: required for `Vec<u8>` to implement `EncodeReturn`
   = note: required for `Vec<u8>` to implement `ConvertReturn`
 note: required by a bound in `send_message`
- --> $WORKSPACE/crates/objc2/src/message.rs
+ --> $WORKSPACE/crates/objc2/src/runtime/message.rs
   |
   |     unsafe fn send_message<A, R>(self, sel: Sel, args: A) -> R
   |               ------------ required by a bound in this associated function
@@ -49,7 +49,7 @@ error[E0277]: the trait bound `Vec<u8>: Encode` is not satisfied
   = note: required for `Vec<u8>` to implement `ConvertArgument`
   = note: required for `(Vec<u8>,)` to implement `MessageArguments`
 note: required by a bound in `send_message`
- --> $WORKSPACE/crates/objc2/src/message.rs
+ --> $WORKSPACE/crates/objc2/src/runtime/message.rs
   |
   |     unsafe fn send_message<A, R>(self, sel: Sel, args: A) -> R
   |               ------------ required by a bound in this associated function
@@ -81,7 +81,7 @@ error[E0277]: the trait bound `(): Encode` is not satisfied
   = note: required for `()` to implement `ConvertArgument`
   = note: required for `((),)` to implement `MessageArguments`
 note: required by a bound in `send_message`
- --> $WORKSPACE/crates/objc2/src/message.rs
+ --> $WORKSPACE/crates/objc2/src/runtime/message.rs
   |
   |     unsafe fn send_message<A, R>(self, sel: Sel, args: A) -> R
   |               ------------ required by a bound in this associated function
