+++ crates/test-ui/ui/msg_send_underspecified_error.stderr
@@ -16,7 +16,7 @@ error[E0283]: type annotations needed
             where T: RefEncode, T: ?Sized;
   = note: required for `*mut *mut _` to implement `Encode`
 note: required by a bound in `__send_message_error`
- --> $WORKSPACE/crates/objc2/src/message.rs
+ --> $WORKSPACE/crates/objc2/src/runtime/message.rs
   |
   |     unsafe fn __send_message_error<A, E>(self, sel: Sel, args: A) -> Result<(), Id<E>>
   |               -------------------- required by a bound in this associated function
